refactor(modules): share ChannelMerger/FourierEmb via merger.py
Move FourierEmb and ChannelMerger out of dance_modules.py into the new
shared braindecode/modules/merger.py so DANCE and BrainModule use one
merger. dance_modules.py now imports them; modules/__init__.py re-exports
ChannelMerger/FourierEmb/positions_from_chs_info from merger. dance.py
drops its local _positions_from_chs_info/_has_valid_locations and imports
the shared positions_from_chs_info/has_valid_locations under the same
private aliases. Positions fed to the merger are unchanged, so parity is
preserved (encoder + forward dense parity OK).
